basic structure for userlist
